add dedupe control component
