+++ frontend/src/stories/features/collections/DedupeWorkflows.ts
@@ -1,7 +1,6 @@
 import { html } from "lit";
 
 import type { DedupeWorkflows } from "@/features/collections/dedupe-workflows/dedupe-workflows";
-import type { Crawl, ListWorkflow } from "@/types/crawler";
 
 import "@/features/collections/dedupe-workflows/dedupe-workflows";
 import "@/features/archived-items/item-dependency-tree";
@@ -14,176 +13,3 @@ export const renderComponent = (props: Partial<RenderProps>) => {
     .workflows=${props.workflows}
   ></btrix-dedupe-workflows>`;
 };
-
-const crawls: Crawl[] = [
-  {
-    id: "item-storybook-id-1",
-    type: "crawl",
-    userid: "item-storybook-userid-1",
-    userName: "Alice",
-    oid: "item-storybook-oid-1",
-    profileid: null,
-    name: "",
-    description: "",
-    started: "2023-10-27T03:20:00Z",
-    finished: "2023-10-27T03:23:00Z",
-    state: "complete",
-    stats: { found: "1", done: "1", size: "32491355" },
-    fileSize: 16257194,
-    fileCount: 1,
-    tags: [],
-    collections: [],
-    crawlExecSeconds: 8,
-    qaCrawlExecSeconds: 0,
-    cid: "item-storybook-cid-1",
-    firstSeed: "https://webrecorder.net/",
-    seedCount: 1,
-    profileName: null,
-    stopping: false,
-    shouldPause: false,
-    manual: true,
-    scale: 1,
-    browserWindows: 4,
-    crawlerChannel: "default",
-    proxyId: null,
-    image: null,
-    qaRunCount: 0,
-    activeQAStats: null,
-    lastQAState: null,
-    lastQAStarted: null,
-    pageCount: 1,
-    requiresCrawls: ["item-storybook-id-2", "item-storybook-id-3"],
-    requiredByCrawls: [],
-    collectionIds: ["workflow-storybook-collId-1"],
-    dedupeCollId: "workflow-storybook-collId-1",
-  },
-  {
-    id: "item-storybook-id-2",
-    type: "crawl",
-    name: "",
-    started: "2023-10-27T03:18:37Z",
-    finished: "2023-10-27T03:20:00Z",
-    state: "complete",
-    stats: { found: "1", done: "1", size: "32491355" },
-    fileSize: 16257194,
-    firstSeed: "https://example.com/",
-    seedCount: 1,
-    pageCount: 1,
-    requiresCrawls: [],
-    requiredByCrawls: [],
-    collectionIds: ["workflow-storybook-collId-1"],
-    dedupeCollId: "workflow-storybook-collId-1",
-  },
-  {
-    id: "item-storybook-id-3",
-    type: "crawl",
-    name: "",
-    started: "2023-10-27T03:15:15Z",
-    finished: "2023-10-27T03:18:37Z",
-    state: "complete",
-    stats: { found: "1", done: "1", size: "32491355" },
-    fileSize: 16257194,
-    firstSeed: "https://example.com/",
-    seedCount: 1,
-    pageCount: 1,
-    requiresCrawls: [],
-    requiredByCrawls: [],
-    collectionIds: ["workflow-storybook-collId-1"],
-    dedupeCollId: "workflow-storybook-collId-1",
-  },
-] as const as Crawl[];
-
-const workflows: ListWorkflow[] = [
-  {
-    name: "",
-    description: null,
-    created: "2025-12-04T22:11:25Z",
-    createdBy: "workflow-storybook-userId-1",
-    modified: "2025-12-04T22:23:18Z",
-    modifiedBy: "workflow-storybook-userId-1",
-    autoAddCollections: ["workflow-storybook-collId-1"],
-    inactive: false,
-    crawlAttemptCount: 5,
-    crawlCount: 5,
-    crawlSuccessfulCount: 5,
-    totalSize: 1336853,
-    lastCrawlId: "workflow-storybook-crawlId-1",
-    lastCrawlStartTime: "2025-12-04T23:01:39Z",
-    lastCrawlTime: "2025-12-04T23:04:46Z",
-    lastCrawlState: "complete",
-    lastCrawlSize: 220956,
-    lastRun: "2025-12-04T23:04:46Z",
-    isCrawlRunning: false,
-    shareable: false,
-    id: "workflow-storybook-id-1",
-    schedule: "",
-    jobType: "custom",
-    config: null,
-    tags: [],
-    crawlTimeout: 0,
-    maxCrawlSize: 0,
-    browserWindows: 1,
-    oid: "workflow-storybook-oid-1",
-    profileid: null,
-    crawlerChannel: "dev-2",
-    proxyId: null,
-    firstSeed: "https://webrecorder.net/",
-    seedCount: 1,
-    dedupeCollId: "workflow-storybook-collId-1",
-    lastCrawlStopping: false,
-    lastCrawlShouldPause: false,
-    lastCrawlPausedAt: null,
-    lastCrawlPausedExpiry: null,
-    profileName: null,
-    createdByName: "Alice",
-    modifiedByName: "Alice",
-    lastStartedByName: "Alice",
-  },
-  {
-    name: "webrecorder.net no dedupe",
-    description: null,
-    created: "2025-12-05T00:08:49Z",
-    createdBy: "workflow-storybook-userId-1",
-    modified: "2025-12-05T00:08:49Z",
-    modifiedBy: "workflow-storybook-userId-1",
-    autoAddCollections: [],
-    inactive: false,
-    crawlAttemptCount: 0,
-    crawlCount: 0,
-    crawlSuccessfulCount: 0,
-    totalSize: 0,
-    lastCrawlId: null,
-    lastCrawlStartTime: null,
-    lastCrawlTime: null,
-    lastCrawlState: null,
-    lastCrawlSize: null,
-    lastRun: null,
-    isCrawlRunning: false,
-    shareable: false,
-    id: "workflow-storybook-id-1",
-    schedule: "",
-    jobType: "custom",
-    config: null,
-    tags: [],
-    crawlTimeout: 0,
-    maxCrawlSize: 0,
-    browserWindows: 1,
-    oid: "workflow-storybook-oid-1",
-    profileid: null,
-    crawlerChannel: "default",
-    proxyId: null,
-    firstSeed: "https://webrecorder.net/",
-    seedCount: 1,
-    dedupeCollId: null,
-    lastCrawlStopping: false,
-    lastCrawlShouldPause: false,
-    lastCrawlPausedAt: null,
-    lastCrawlPausedExpiry: null,
-    profileName: null,
-    createdByName: "Alice",
-    modifiedByName: "Alice",
-    lastStartedByName: null,
-  },
-] as const as ListWorkflow[];
-
-export const data = { crawls, workflows };
